// src/models/navigation.d.ts
export type RootStackParamList = {
  Root: undefined;
  SplashScreen: undefined;
  AuthStack: undefined;
  AppStack: undefined;
  HomeStack: undefined;
  HomeTab: undefined;
  ProjectTab: undefined;
  P2PTab: undefined;
  DefiTab: undefined;
  PartnerTab: undefined;

  //Screen
  Splash: undefined;
  Register: undefined;
  Login: undefined;
  VerifySuccessScreen: undefined;
  RegisterOnboardScreen: undefined;
  HomeScreen: undefined;
  CountryScreen: undefined;
  SettingScreen: undefined;
  LanguageScreen: undefined;
  MyProfileScreen: undefined;
  UpdateEmailScreen: undefined;
  ChangePasswordScreen: undefined;
  TouchIDScreen: undefined;
  IDCardScreen: undefined;
  RecordScreen: undefined;
  KYCPendingScreen: undefined;
  KYCSuccessScreen: undefined;
  RejectedScreen: undefined;
  DigitalAssetScreen: undefined;
  ReceiveScreen: undefined;
  ReceiveDigitalScreen: undefined;
  DigitalDetailScreen: undefined;
  TransactionHistoryScreen: undefined;
  SendScreen: undefined;
  SendDigitalScreen: undefined;
  ScanQRCodeScreen: undefined;
  ToAddressScreen: undefined;
  ConfirmSendScreen: undefined;
  ForgotPasswordScreen: undefined;
  VerifyOTPScreen: undefined;

  //KYC
  KYCOnboardingScreen: undefined;
  PersonalInformationScreen: undefined;
};